Fix more warnings.

+++ Projects/Src/Compression.SevenZipDLLDecoder.pas
@@ -661,7 +661,7 @@ procedure TArchiveExtractBaseCallback.HandleResult;
     if Res = E_OUTOFMEMORY then
       SevenZipError(Win32ErrorString(DWORD(E_OUTOFMEMORY)))
     else
-      SevenZipWin32Error('Extract', FResult.Res);
+      SevenZipWin32Error('Extract', DWORD(FResult.Res));
   end;
 
 begin
